build: update eslint-config-prettier 4.x
