+++ package.json
@@ -41,14 +41,14 @@
     "eslint": "^5.5.0",
     "eslint-config-airbnb": "^17.0.0",
     "eslint-config-airbnb-base": "^13.0.0",
-    "eslint-config-prettier": "^3.0.0",
+    "eslint-config-prettier": "^4.2.0",
     "eslint-plugin-import": "^2.14.0",
     "eslint-plugin-jest": "^22.1.3",
     "eslint-plugin-jsx-a11y": "^6.1.1",
     "eslint-plugin-promise": "^4.0.0",
     "eslint-plugin-react": "^7.11.1",
-    "eslint-plugin-unicorn": "^8.0.2",
-    "eslint-plugin-react-hooks": "^1.2.0"
+    "eslint-plugin-react-hooks": "^1.2.0",
+    "eslint-plugin-unicorn": "^8.0.2"
   },
   "devDependencies": {
     "@commitlint/cli": "^7.1.2",
